Render to texture applied

+++ BruteForceEngine/TextureDX12.cpp
@@ -15,7 +15,7 @@ namespace BruteForce
             shaderResourceViewDesc.ViewDimension = D3D12_SRV_DIMENSION_TEXTURE2D;
             shaderResourceViewDesc.Shader4ComponentMapping = D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING;
             shaderResourceViewDesc.Format = Format;
-            shaderResourceViewDesc.Texture2D.MipLevels = m_Mips;
+            shaderResourceViewDesc.Texture2D.MipLevels = static_cast<UINT>(m_Mips);
             shaderResourceViewDesc.Texture2D.MostDetailedMip = 0;
             shaderResourceViewDesc.Texture2D.ResourceMinLODClamp = 0.0f;
             device->CreateShaderResourceView(image.Get(), &shaderResourceViewDesc, descriptor_handle);
